% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichDot.R
\name{as.enrichdat}
\alias{as.enrichdat}
\title{Adjust dataframe for enrichment plot
make sure colnames have Description, Count, FoldEnrich/GeneRatio, pvalue/qvalue/p.adjust}
\usage{
as.enrichdat(enrich_df)
}
\arguments{
\item{enrich_df}{dataframe of enrichment analysis result .}
}
\value{
A `data.frame`.
}
\description{
Adjust dataframe for enrichment plot
make sure colnames have Description, Count, FoldEnrich/GeneRatio, pvalue/qvalue/p.adjust
}
