% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{msig_org}
\alias{msig_category}
\alias{biocOrg_name}
\alias{keggOrg_name}
\title{Datasets
msig_species contains msigdb species information}
\description{
Datasets
msig_species contains msigdb species information

Datasets
msig_category contains msigdb category information

Datasets
biocOrg_name contains organism name of bioconductor

Datasets
keggOrg_name contains organism name of KEGG
https://www.genome.jp/kegg/catalog/org_list.html
}
\examples{
data(msig_org)
str(msig_org)
data(msig_category)
str(msig_category)
data(biocOrg_name)
str(biocOrg_name)
data(keggOrg_name)
str(keggOrg_name)
}
\keyword{datasets}
