% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process.R
\name{expo_sheet}
\alias{expo_sheet}
\title{export result into different sheets}
\usage{
expo_sheet(wb, sheet_dat, sheet_name)
}
\arguments{
\item{wb}{worksheet from `createWorkbook()`.}

\item{sheet_dat}{dataframe added to sheet.}

\item{sheet_name}{name of added dataframe.}
}
\value{
a worksheet including many dataframes.
}
\description{
export result into different sheets
}
\examples{
\donttest{
expo_sheet(wb, sheet_dat =  mtcars, sheet_name = 'mtcars')
}
}
