% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoSheet.R
\name{expoSheet}
\alias{expoSheet}
\title{Export result into different Excel sheets}
\usage{
expoSheet(wb, sheet_dat, sheet_name)
}
\arguments{
\item{wb}{workbook created from `createWorkbook`.}

\item{sheet_dat}{`data.frame` to export.}

\item{sheet_name}{sheet name.}
}
\value{
An Excel file.
}
\description{
Export result into different Excel sheets
}
\examples{
\donttest{
library(openxlsx)
wb <- createWorkbook()
wb <- expoSheet(wb, sheet_dat =  mtcars, sheet_name = 'mtcars')
saveWorkbook(wb, paste0(tempdir(),'/test.xlsx'), overwrite = TRUE)
}

}
