% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGO.R
\name{genGO}
\alias{genGO}
\title{Gene enrichment of GO analysis}
\usage{
genGO(
  id,
  org,
  ont,
  use_symbol = TRUE,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.1,
  minGSSize = 10,
  maxGSSize = 500,
  universe,
  ...
)
}
\arguments{
\item{id}{a gene vector which can be entrez, ensembl or symbol.}

\item{org}{organism name from `biocOrg_name`.}

\item{ont}{One of "bp", "mf", and "cc" subontologies, or "all" for all three.}

\item{use_symbol}{logical to output as gene symbol, default is TRUE.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{pvalueCutoff}{adjusted pvalue cutoff, default is 0.05.}

\item{qvalueCutoff}{adjusted pvalue cutoff, default is 0.1.}

\item{minGSSize}{minimal size of each geneSet for analyzing, default is 10.}

\item{maxGSSize}{maximal size of each geneSet for analyzing, default is 500.}

\item{universe}{background genes. If missing, the orgdb all gene list will be used as background.}
}
\value{
a dataframe of gene info.
}
\description{
Gene enrichment of GO analysis
}
\examples{
\dontrun{
data(geneList, package="DOSE")
id = names(geneList)[1:100]
ego <- genGO(id, org = 'human',ont = 'mf',pvalueCutoff = 0.05,qvalueCutoff = 0.1 ,use_symbol = T)
head(ego)
}
}
