% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInfo.R
\name{genInfo}
\alias{genInfo}
\title{Get gene related information}
\usage{
genInfo(id, org)
}
\arguments{
\item{id}{Gene id.}

\item{org}{Species name from `biocOrg_name`.}
}
\value{
A `data.frame`.
}
\description{
Get gene related information
}
\examples{
x = genInfo(id= c("Cyp2c23","Fhit","Gal3st2b","Trp53","Tp53"), org = 'mm')
head(x)
}
