% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genKEGG.R
\name{genKEGG}
\alias{genKEGG}
\title{Gene enrichment of KEGG analysis}
\usage{
genKEGG(
  id,
  org,
  use_symbol = TRUE,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.1,
  minGSSize = 10,
  maxGSSize = 500,
  universe,
  ...
)
}
\arguments{
\item{id}{a vector of entrez gene.}

\item{org}{character of organism name which can test by `mapKeggOrg()`.}

\item{use_symbol}{logical to output as gene symbol, default is TRUE.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{pvalueCutoff}{numberic of adjusted pvalue cutoff, default is 0.05.}

\item{qvalueCutoff}{numberic of adjusted pvalue cutoff, default is 0.1.}

\item{minGSSize}{numberic of minimal size of each geneSet for analyzing, default is 10.}

\item{maxGSSize}{numberic of maximal size of each geneSet for analyzing, default is 500.}

\item{universe}{background genes. If missing, the orgdb all gene list will be used as background.}
}
\value{
a dataframe of gene info.
}
\description{
Gene enrichment of KEGG analysis
}
\examples{
\dontrun{
data(geneList, package="DOSE")
id = names(geneList)[1:100]
keg <- genKEGG(id, org = 'human')
head(keg)
}
}
