% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genORA.R
\name{genORA}
\alias{genORA}
\title{Gene ORA enrichment analysis for other data sets}
\usage{
genORA(
  id,
  type = c("do", "ncg", "dgn", "wiki"),
  org = "human",
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  minGSSize = 10,
  maxGSSize = 500,
  group_list = NULL,
  universe
)
}
\arguments{
\item{id}{A vector of gene id which can be entrezid, ensembl, symbol or uniprot.}

\item{type}{ORA analyse type, choose one from "do" (Disease Ontology), "ncg" (Network of Cancer Gene),
"dgn" (DisGeNET) and "wiki" (WikiPathways)}

\item{org}{human as default}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH",
"BY", "fdr", "none".}

\item{pvalueCutoff}{Numberic of pvalue cutoff, default is 0.05.}

\item{qvalueCutoff}{Numberic of adjusted pvalue cutoff, default is 0.05.}

\item{minGSSize}{Numberic of minimal size of each geneSet for analyzing,
default is 10.}

\item{maxGSSize}{Numberic of maximal size of each geneSet for analyzing,
default is 500.}

\item{group_list}{A list of gene id groups, default is NULL.}

\item{universe}{Background genes. If missing, the orgdb all gene list will be
used as background.}
}
\value{
A `data.frame`.
}
\description{
Gene ORA enrichment analysis for other data sets
}
\examples{
\dontrun{
# only gene ids
data(geneList, package = "genekitr")
id <- names(geneList)[abs(geneList) > 1]
# Disease ontology (DO)
edo <- genORA(id, type = "do")
# Network of Cancer Gene (NCG)
encg <- genORA(names(geneList), type = "ncg")
# DisGeNET (DGN)
edgn <- genORA(id, type = "dgn", pvalueCutoff = 0.01, qvalueCutoff = 0.01)
# WikiPathways (wiki)
ewiki <- genORA(id, type = "wiki", org = "human")

# gene id with groups
id <- c(head(names(geneList), 100), tail(names(geneList), 100))
group <- list(
  group1 = c(rep("up", 100), rep("down", 100)),
  group2 = c(rep("A", 130), rep("B", 70))
)
gedo <- genORA(id, type = "do",group_list = group)

}
}
