% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPubmed.R
\name{genPubmed}
\alias{genPubmed}
\title{Get gene related pubmed paper record}
\usage{
genPubmed(id, keywords, field = "ALL", ...)
}
\arguments{
\item{id}{gene "symbol".}

\item{keywords}{species "mm" or "hs".}

\item{field}{pubmed field from `showNCBI('pubmed')`, default is "ALL".}
}
\value{
a dataframe of pubmed paper records of genes.
}
\description{
Get gene related pubmed paper record
}
\examples{
\dontrun{
x = genPubmed(id = c("Cyp2c23","Fhit","Gal3st2b","Gbp4"), keywords = 'stem cell', field = 'tiab')
}

}
