% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analyse.R
\name{getMsigdb}
\alias{getMsigdb}
\title{Get Msigdb database term and gene information}
\usage{
getMsigdb(
  org,
  category = c("C1", "C2", "C3", "C4", "C5", "C6", "C7", "C8", "H"),
  subcategory = NULL,
  ...
)
}
\arguments{
\item{org}{organism name from `msigdb_org_data()`.}

\item{category}{MSigDB collection abbreviation, C1 to C8 and H.}

\item{subcategory}{MSigDB sub-collection abbreviation, such as REACTOME or BP.}
}
\value{
a dataframe of 3 columns with term, entrez and symbol name.
}
\description{
Get Msigdb database term and gene information
}
\examples{
\donttest{
msigdb <- getMsigdb(org='human', category='C5',subcategory='GO:CC')
}
}
