% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCP.R
\name{importCP}
\alias{importCP}
\title{Import 'clusterProfiler' result}
\usage{
importCP(object, type = c("go", "gsea", "other"))
}
\arguments{
\item{object}{clusterProfiler object.}

\item{type}{object type from "go", "gsea" and "other". "other" includes ORA (over-representation analysis) of KEGG, DOSE,...}
}
\value{
`data.frame`
}
\description{
Import 'clusterProfiler' result
}
