% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPanther.R
\name{importPanther}
\alias{importPanther}
\title{Import 'Panther' web result}
\usage{
importPanther(panther_file)
}
\arguments{
\item{panther_file}{Panther result file.}
}
\value{
`data.frame`
}
\description{
Import 'Panther' web result
}
