% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importShinygo.R
\name{importShinygo}
\alias{importShinygo}
\title{Import 'shinyGO' web result}
\usage{
importShinygo(shinygo_file)
}
\arguments{
\item{shinygo_file}{ShinyGO result file.}
}
\value{
`data.frame`
}
\description{
Import 'shinyGO' web result
}
