% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_theme}
\alias{plot_theme}
\title{plot_theme}
\usage{
plot_theme(
  main_text_size = 14,
  legend_text_size = 10,
  font_type = "Arial",
  border_thick = 1,
  ...
)
}
\arguments{
\item{main_text_size}{numeric, main text font size.}

\item{legend_text_size}{numeric, legend text font size.}

\item{font_type}{font family.}

\item{border_thick}{numeric, border thickness in mm.}
}
\value{
ggplot theme
}
\description{
ggplot theme
}
\examples{
library(ggplot2)
ggplot(mtcars,aes(x=mpg)) + geom_histogram(binwidth=5) + plot_theme()
}
