% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichDot.R
\name{plotEnrichDot}
\alias{plotEnrichDot}
\title{Dotplot for enrichment analysis}
\usage{
plotEnrichDot(
  enrich_df,
  xlab_type = c("FoldEnrich", "GeneRatio", "Count"),
  legend_by = c("p.adjust", "pvalue", "qvalue"),
  border_thick = 1,
  remove_grid = FALSE,
  remove_text = FALSE,
  remove_legend = FALSE,
  low_color = "red",
  high_color = "blue",
  font_type = "Arial",
  show_item = 10,
  main_text_size = 10,
  legend_text_size = 8,
  wrap_width = NULL,
  ...
)
}
\arguments{
\item{enrich_df}{dataframe of enrichment analysis result .}

\item{xlab_type}{x-axis label type, one of 'GeneRatio','Count','FoldEnrich'.}

\item{legend_by}{stats legend type, one of "pvalue", "p.adjust", "qvalue".}

\item{border_thick}{numeric, border thickness in mm.}

\item{remove_grid}{logical, remove legend, default is FALSE.}

\item{remove_text}{logical, remove all text, default is FALSE.}

\item{font_type}{character, specify the plot text font family, example "Times New Roman", "Arial".}

\item{show_item}{numeric, select top N rows to show.}

\item{main_text_size}{numeric, specify the plot text size.}

\item{wrap_width}{numeric, wrap text longer than this number.}
}
\value{
ggplot object
}
\description{
Dotplot for enrichment analysis
}
\examples{
\dontrun{
data(geneList, package="DOSE")
id = names(geneList)[1:100]
ego = genGO(id, org = 'human',ont = 'mf',pvalueCutoff = 0.05,qvalueCutoff = 0.1 ,use_symbol = T)
ego = as.enrichDat(ego)
plotEnrichDot(ego)
}
}
\author{
Yunze Liu
}
