% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process.R
\name{showNCBI}
\alias{showNCBI}
\title{Show NCBI database searchable name}
\usage{
showNCBI(db = "pubmed")
}
\arguments{
\item{db}{a character vector. Can be "pubmed" or one or more of `rentrez::entrez_dbs()` result.}
}
\value{
a dataframe including search keyword and information.
}
\description{
Show NCBI database searchable name
}
\examples{
\donttest{
showNCBI("pubmed")
}
}
