% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGO.R
\name{simGO}
\alias{simGO}
\title{Simplitify GO enrichment result}
\usage{
simGO(
  enrich_df,
  sim_method = c("JC", "Resnik", "Lin", "Rel", "Jiang", "Wang"),
  org = NULL,
  ont = NULL
)
}
\arguments{
\item{enrich_df}{Enrichment analysis `data.frame` result.}

\item{sim_method}{Method of calculating the similarity between nodes, one of one of "Resnik",
"Lin", "Rel", "Jiang" , "Wang" and "JC" (Jaccard similarity coefficient) methods.}

\item{org}{Organism name from `biocOrg_name`.}

\item{ont}{One of "bp", "mf", and "cc".}
}
\value{
A `data.frame` contains simplified GO terms.
}
\description{
Simplitify GO enrichment result
}
