% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transID.R
\name{transId}
\alias{transId}
\title{transform gene id among symbol, entrezid and ensembl}
\usage{
transId(id, trans_to, org, return_dat = FALSE)
}
\arguments{
\item{id}{character of gene ids.}

\item{trans_to}{transform to which type, one of symbol, entrezid and ensembl.}

\item{org}{organism name from `biocOrg_name`.}

\item{return_dat}{logical, default is `TRUE`, return a dataframe of transformed ids.}
}
\value{
a character or dataframe of transformed ids.
}
\description{
transform gene id among symbol, entrezid and ensembl
}
\examples{
\donttest{
msigdb <- getMsigdb(org='human', category='C5',subcategory='GO:CC')
}
}
