% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transProbe.R
\name{transProbe}
\alias{transProbe}
\title{Transform probe id to symbol, entrezid,  ensembl or uniprot.}
\usage{
transProbe(id, transTo, org = "human", platform = NULL)
}
\arguments{
\item{id}{probe ids.}

\item{transTo}{Transform to what type. User could select one or more from
"symbol", "entrez", "ensembl" or "uniprot."}

\item{org}{'human'.}

\item{platform}{Probe platform. If NULL, program will detect automatically.}
}
\value{
data frame, first column is probe id and others are converted id.
}
\description{
Transform probe id to symbol, entrezid,  ensembl or uniprot.
}
