% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asEnrichdat.R
\name{as.enrichdat}
\alias{as.enrichdat}
\title{Modify dataframe for enrichment plot}
\usage{
as.enrichdat(enrich_df)
}
\arguments{
\item{enrich_df}{Enrichment analysis `data.frame` result.}
}
\value{
`data.frame`
}
\description{
To make sure colname contains Description, Count, FoldEnrich/GeneRatio, pvalue/qvalue/p.adjust
}
