% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asEnrichdat.R
\name{as.enrichdat}
\alias{as.enrichdat}
\title{Adjust dataframe for enrichment plot}
\usage{
as.enrichdat(enrich_df)
}
\arguments{
\item{enrich_df}{dataframe of enrichment analysis result .}
}
\value{
A `data.frame`.
}
\description{
make sure colname contains Description, Count, FoldEnrich/GeneRatio, pvalue/qvalue/p.adjust
}
