% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{geneList}
\alias{msig_org}
\alias{msig_category}
\alias{biocOrg_name}
\alias{keggOrg_name}
\alias{ensOrg_name}
\title{Datasets
geneList entrez gene list with decreasing fold change value}
\description{
Datasets
geneList entrez gene list with decreasing fold change value

Datasets
msig_species contains msigdb species information

Datasets
msig_category contains msigdb category information

Datasets
biocOrg_name contains organism name of bioconductor

Datasets
keggOrg_name contains organism name of KEGG
https://www.genome.jp/kegg/catalog/org_list.html

Datasets
ensOrg_name contains organism name of ensembl
}
\keyword{datasets}
