% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoSheet.R
\name{expo_sheet}
\alias{expo_sheet}
\title{Export result into different Excel sheets}
\usage{
expo_sheet(wb, sheet_dat, sheet_name)
}
\arguments{
\item{wb}{workbook created from `createWorkbook`.}

\item{sheet_dat}{`data.frame` to export.}

\item{sheet_name}{sheet name.}
}
\value{
An Excel file.
}
\description{
Export result into different Excel sheets
}
\examples{
\dontrun{
library(openxlsx)
wb <- createWorkbook()
wb <- expo_sheet(wb, sheet_dat =  mtcars, sheet_name = 'mtcars')
saveWorkbook(wb, "./test.xlsx", overwrite = TRUE)
}

}
