% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoSheet.R
\name{expoSheet}
\alias{expoSheet}
\title{Export list of datasets into different Excel sheets}
\usage{
expoSheet(
  dat_list,
  name_list,
  filename = NULL,
  dir = tempdir(),
  overwrite = TRUE
)
}
\arguments{
\item{dat_list}{List of datasets.}

\item{name_list}{List of data names.}

\item{filename}{A character string naming an xlsx file.}

\item{dir}{A character string naming output directory.}

\item{overwrite}{If TRUE, overwrite any existing file.}
}
\value{
An Excel file.
}
\description{
Export list of datasets into different Excel sheets
}
\examples{
\donttest{
library(openxlsx)
expoSheet(dat_list =  list(mtcars,ToothGrowth), name_list = list('mtcars','tooth'),
  filename = 'test.xlsx', dir = tempdir())
}

}
