% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGSEA.R
\name{genGSEA}
\alias{genGSEA}
\title{GSEA for a genelist with decreasing logFC value}
\usage{
genGSEA(
  genelist,
  org,
  category = c("C1", "C2", "C3", "C4", "C5", "C6", "C7", "C8", "H"),
  subcategory = NULL,
  use_symbol = TRUE,
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  ...
)
}
\arguments{
\item{genelist}{order ranked genelist in decreasing order, gene can be entrez or symbol.}

\item{org}{organism name from `msigdb_org_data()`.}

\item{category}{MSigDB collection abbreviation, C1 to C8 and H.}

\item{subcategory}{MSigDB sub-collection abbreviation, such as REACTOME or BP.}

\item{use_symbol}{logical to output as gene symbol, default is TRUE.}

\item{minGSSize}{minimal size of each geneSet for analyzing, default is 10.}

\item{maxGSSize}{maximal size of each geneSet for analyzing, default is 500.}

\item{pvalueCutoff}{adjusted pvalue cutoff, default is 0.05.}
}
\value{
a dataframe of gene info.
}
\description{
GSEA for a genelist with decreasing logFC value
}
\examples{
\dontrun{
data(geneList, package="DOSE")
genGSEA(genelist = geneList,org = 'human', category='C3',subcategory = 'TFT:GTRD',use_symbol = F)
}
}
