% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInfo.R
\name{genInfo}
\alias{genInfo}
\title{Get gene info using org.db}
\usage{
genInfo(id, org, ...)
}
\arguments{
\item{id}{gene id.}

\item{org}{species name from `biocOrg_name()`.}
}
\value{
a dataframe of gene info.
}
\description{
Get gene info using org.db
}
\examples{
\donttest{
x = genInfo(id= c("Cyp2c23","Fhit","Gal3st2b","Trp53","Tp53"), org = 'mm')
}
}
