% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInfo.R
\name{genInfo}
\alias{genInfo}
\title{Get gene related information}
\usage{
genInfo(id, org, simple = FALSE)
}
\arguments{
\item{id}{Gene id (symbol, ensembl or entrez id) or uniprot id.}

\item{org}{Species name from `biocOrg_name()`, both full name and short name are fine.}

\item{simple}{Logical to keep only one matched ID, default is FALSE.}
}
\value{
A `data.frame`.
}
\description{
Get gene related information
}
\examples{
x <- genInfo(id = c(
  "MCM10", "CDC20", "S100A9", "MMP1", "BCC7",
  "FAKEID", "TP53", "HBD", "TP53", "NUDT10"
), org = "hg", simple = FALSE)
head(x)
}
