% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genKEGG.R
\name{genKEGG}
\alias{genKEGG}
\title{Gene ORA-KEGG enrichment analysis}
\usage{
genKEGG(
  id,
  org,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.15,
  minGSSize = 10,
  maxGSSize = 500,
  group_list = NULL,
  universe,
  ...
)
}
\arguments{
\item{id}{A vector of gene id which can be entrezid, ensembl, symbol or uniprot.}

\item{org}{KEGG organism name from `keggOrg_name`.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH",
"BY", "fdr", "none".}

\item{pvalueCutoff}{Numberic of pvalue cutoff, default is 0.05.}

\item{qvalueCutoff}{Numberic of adjusted pvalue cutoff, default is 0.05.}

\item{minGSSize}{Numberic of minimal size of each geneSet for analyzing,
default is 10.}

\item{maxGSSize}{Numberic of maximal size of each geneSet for analyzing,
default is 500.}

\item{group_list}{A list of gene id groups, default is NULL.}

\item{universe}{Background genes. If missing, the orgdb all gene list will be
used as background.}

\item{...}{Other argument to `enrichKEGG` function}
}
\value{
A `data.frame`.
}
\description{
Gene ORA-KEGG enrichment analysis
}
\examples{
\dontrun{
# only gene ids
data(geneList, package = "genekitr")
id <- names(geneList)[abs(geneList) > 1]
keg <- genKEGG(id, org = "human")

# gene id with groups
id <- c(head(names(geneList), 100), tail(names(geneList), 100))
group <- list(
  group1 = c(rep("up", 100), rep("down", 100)),
  group2 = c(rep("A", 130), rep("B", 70))
)
gkeg <- genKEGG(id,
  group_list = group,
  org = "human", pvalueCutoff = 0.05,
  qvalueCutoff = 0.05
)
}

}
