% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPubmed.R
\name{genPubmed}
\alias{genPubmed}
\title{Gene related pubmed paper records}
\usage{
genPubmed(id, keywords, field = "ALL")
}
\arguments{
\item{id}{Gene "symbol".}

\item{keywords}{species "mm" or "hs".}

\item{field}{pubmed field from `showNCBI('pubmed')`, default is "ALL".}
}
\value{
A `data.frame`.
}
\description{
Gene related pubmed paper records
}
\examples{
\donttest{
x = genPubmed(id = c("Cyp2c23","Fhit","Gal3st2b","Gbp4"),
  keywords = 'stem cell', field = 'tiab')
}
}
