% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_theme}
\alias{plot_theme}
\title{ggplot theme}
\usage{
plot_theme(
  main_text_size = 14,
  legend_text_size = 10,
  font_type = "sans",
  border_thick = 1
)
}
\arguments{
\item{main_text_size}{Numeric, main text size}

\item{legend_text_size}{Numeric, legend text size}

\item{font_type}{Font family}

\item{border_thick}{Numeric, border thickness}
}
\value{
ggplot theme
}
\description{
Change ggplot text, font and border
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=wt, y=mpg))+ geom_point()+ plot_theme(font_type = 'Times')
}
