% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichDot.R
\name{plotEnrichDot}
\alias{plotEnrichDot}
\title{Dotplot for GO and KEGG enrichment analysis}
\usage{
plotEnrichDot(
  enrich_df,
  xlab_type = c("FoldEnrich", "GeneRatio", "Count"),
  legend_by = c("p.adjust", "pvalue", "qvalue"),
  border_thick = 1,
  remove_grid = FALSE,
  remove_text = FALSE,
  remove_legend = FALSE,
  low_color = "red",
  high_color = "blue",
  font_type = "Arial",
  show_item = 10,
  main_text_size = 10,
  legend_text_size = 8,
  wrap_width = NULL
)
}
\arguments{
\item{enrich_df}{`data.frame` of enrichment analysis result .}

\item{xlab_type}{X-axis label type, one of 'GeneRatio','Count','FoldEnrich'.}

\item{legend_by}{Stats legend type, one of "pvalue", "p.adjust", "qvalue".}

\item{border_thick}{Numeric, border thickness in mm.}

\item{remove_grid}{Logical, remove background grid lines, default is FALSE.}

\item{remove_text}{Logical, remove all text, default is FALSE.}

\item{remove_legend}{Logical, remove legend, default is FALSE.}

\item{low_color}{Legend color for low pvalue or qvalue, default is "red".}

\item{high_color}{Legend color for high pvalue or qvalue, default is "blue".}

\item{font_type}{Character, specify the plot text font family, example "Times
New Roman", "Arial".}

\item{show_item}{Numeric, select top N rows to show, default is 10.}

\item{main_text_size}{Numeric, specify the main text size, default is 10.}

\item{legend_text_size}{Numeric, specify the legend text size, default is 8.}

\item{wrap_width}{Numeric, wrap text if longer than this number, default is NULL.}
}
\value{
A ggplot object
}
\description{
Dotplot for GO and KEGG enrichment analysis
}
\examples{
\dontrun{
data(geneList, package="DOSE")
id = names(geneList)[1:100]
ego = genGO(id, org = 'human',ont = 'mf',pvalueCutoff = 0.05,
  qvalueCutoff = 0.1 ,use_symbol = FALSE)
ego = as.enrichDat(ego)
plotEnrichDot(ego)
}
}
