% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichDot.R
\name{plotEnrichDot}
\alias{plotEnrichDot}
\title{Dotplot for GO and KEGG enrichment analysis}
\usage{
plotEnrichDot(
  enrich_df,
  xlab_type = c("FoldEnrich", "GeneRatio", "Count"),
  legend_type = c("p.adjust", "pvalue", "qvalue"),
  low_color = "red",
  high_color = "blue",
  show_item = 10,
  xlim_left = 0,
  xlim_right = NA,
  wrap_length = NULL,
  ...
)
}
\arguments{
\item{enrich_df}{`data.frame` of enrichment analysis result .}

\item{xlab_type}{X-axis label type, one of 'GeneRatio','Count','FoldEnrich'.}

\item{legend_type}{Stats legend type, one of "pvalue", "p.adjust", "qvalue".}

\item{low_color}{Legend color for low pvalue or qvalue, default is "red".}

\item{high_color}{Legend color for high pvalue or qvalue, default is "blue".}

\item{show_item}{Numeric, select top N rows to show, default is 10.}

\item{xlim_left}{X-axis left limit, default is 0.}

\item{xlim_right}{X-axis right limit, default is NA.}

\item{wrap_length}{Numeric, wrap text if longer than this length, default is NULL.}

\item{...}{other arguments transfer to `plot_theme` function}
}
\value{
A ggplot object
}
\description{
Dotplot for GO and KEGG enrichment analysis
}
\examples{
\dontrun{
library(ggplot2)
data(geneList, package = "DOSE")
id <- names(geneList)[1:100]
ego <- genGO(id,
  org = "human", ont = "mf", pvalueCutoff = 0.05,
  qvalueCutoff = 0.1, use_symbol = FALSE
)
plotEnrichDot(ego,remove_grid = T, main_text_size = 8,
  legend_text_size = 6,border_thick = 1.5)
}

}
