% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{plotVenn}
\alias{plotVenn}
\title{Venn plot for list of genes
If venn list length is over 4, use UpSet plot}
\usage{
plotVenn(
  venn_list,
  color = NULL,
  alpha_degree = 0.3,
  border_thick = 1,
  text_size = 1,
  remove_grid = FALSE,
  use_venn = TRUE,
  ...
)
}
\arguments{
\item{venn_list}{a group list of genes or others to plot venn}

\item{color}{colors for venn_list, default is NULL.}

\item{alpha_degree}{alpha transparency of each circle's area, default is 0.3.}

\item{border_thick}{thickness of each circle, default is 1.}

\item{text_size}{text size, default is 1.}

\item{remove_grid}{logical, remove circle or grid lines, default is `FALSE`.}

\item{use_venn}{logical, use venn to plot, default is `TRUE`, the other option is upsetplot for large list.}
}
\value{
ggplot object
}
\description{
Venn plot for list of genes
If venn list length is over 4, use UpSet plot
}
\examples{
\dontrun{
set1 <- paste(rep("gene" , 100) , sample(c(1:1000) , 100 , replace=F) , sep="")
set2 <- paste(rep("gene" , 100) , sample(c(1:1000) , 100 , replace=F) , sep="")
set3 <- paste(rep("gene" , 100) , sample(c(1:1000) , 100 , replace=F) , sep="")
set4 <- paste(rep("gene" , 100) , sample(c(1:1000) , 100 , replace=F) , sep="")
set5 <- paste(rep("gene" , 100) , sample(c(1:1000) , 100 , replace=F) , sep="")
sm_gene_list = list(gset1 = set1, gset2 = set2, gset3 = set3)
la_gene_list = list(gset1 = set1, gset2 = set2, gset3 = set3, gset4 = set4, gset5 = set5 )
plotVenn(sm_gene_list,text_size = 1.5,alpha_degree = 1,remove_grid = T,color = ggsci::pal_lancet()(3))
plotVenn(la_gene_list,text_size = 15,alpha_degree = 0.2,border_thick = 2,
remove_grid = T, use_venn = F)

}
}
