% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGO.R
\name{simGO}
\alias{simGO}
\title{Simplify GO enrichment result}
\usage{
simGO(
  enrich_df,
  sim_method = c("Resnik", "Lin", "Rel", "Jiang", "Wang"),
  org = NULL,
  ont = NULL
)
}
\arguments{
\item{enrich_df}{GO enrichment analysis of `genORA()` result.}

\item{sim_method}{Method of calculating the similarity between nodes, one of one of "Resnik",
"Lin", "Rel", "Jiang" , "Wang" methods.}

\item{org}{Organism name from `biocOrg_name`.}

\item{ont}{One of "bp", "mf", and "cc".}
}
\value{
A `data.frame` contains simplified GO terms.
}
\description{
The Gene Ontology (GO) is a major bioinformatics initiative to unify the representation of gene and gene product
attributes across all species.
}
