% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transID.R
\name{transId}
\alias{transId}
\title{Transform gene id among symbol, entrezid,  ensembl and uniprot}
\usage{
transId(id, trans_to, org, keep_unique = TRUE)
}
\arguments{
\item{id}{Character of gene ids.}

\item{trans_to}{Transform to which type, one of "symbol", "entrezid",
"ensembl" and "uniprot."}

\item{org}{Organism name from `biocOrg_name`.}

\item{keep_unique}{Logical to keep output id is unique, default is `TRUE`.}
}
\value{
A character of transformed ids.
}
\description{
Transform gene id among symbol, entrezid,  ensembl and uniprot
}
\examples{
transId(id= c("Cyp2c23","Fhit","Gal3st2b","Trp53","Tp53"),
  trans_to = 'ensembl', org = 'mouse')
}
