% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transID.R
\name{transId}
\alias{transId}
\title{Transform gene id among symbol, entrezid,  ensembl and uniprot}
\usage{
transId(id, trans_to, org, simple = TRUE)
}
\arguments{
\item{id}{Gene ids.}

\item{trans_to}{Transform to which type, one of "symbol", "entrezid",
"ensembl" and "uniprot."}

\item{org}{Organism name from `biocOrg_name()`, both full name and short name are fine.}

\item{simple}{Logical to keep only one matched ID, default is FALSE.}
}
\value{
A character of transformed ids.
}
\description{
Transform gene id among symbol, entrezid,  ensembl and uniprot
}
\examples{
transId(
  id = c("Cyp2c23", "Fhit", "Gal3st2b", "Trp53", "Tp53"),
  trans_to = "ensembl", org = "mouse", simple = TRUE
)
# input id contains duplicates,fake id and one-to-many match id
transId(
id = c("MMD2", "HBD", "TP53", "RNR1", "TEC", "BCC7", "FAKEID", "TP53"),
trans_to = "entrez", org = "hg", simple = FALSE
)
}
